requirements.txt
mysql-connector-python
argparse
discord.py==1.7.3
selenium
enquiries
rich
requests
pyvirtualdisplay
pillow
EasyProcess
pyotp